tests/wpt/tests/css/css-animations/parsing/animation-trigger-range-end-invalid.html

<!DOCTYPE html>
<!-- TODO(crbug.com/390314945): Replace with spec link when the spec lands. -->
<link rel="help" href="https://github.com/w3c/csswg-drafts/issues/8942">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<script>
test_invalid_value("animation-trigger-range-end", "infinite");
test_invalid_value("animation-trigger-range-end", "1s 2s");
test_invalid_value("animation-trigger-range-end", "1s / 2s");

test_invalid_value("animation-trigger-range-end", "peek 50%");
test_invalid_value("animation-trigger-range-end", "50% contain");
test_invalid_value("animation-trigger-range-end", "50% cover");
test_invalid_value("animation-trigger-range-end", "50% entry");
test_invalid_value("animation-trigger-range-end", "50% enter");
test_invalid_value("animation-trigger-range-end", "50% exit");
test_invalid_value("animation-trigger-range-end", "contain contain");
test_invalid_value("animation-trigger-range-end", "none");
test_invalid_value("animation-trigger-range-end", "cover 50% enter 50%");
</script>
